% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{read_counts}
\alias{read_counts}
\title{GSE151923: cortex from 6-month-old wildtype C57BL/6 mice}
\format{
\subsection{\code{read_counts}}{

A tibble with 1,000 rows and 13 columns.
Each column is one mouse sample from GSE151923.
Each row is a ENSEMBL mouse gene/transcript with the quantified values for each sample.
Gene identifiers are in the first column.
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE151923}
}
\usage{
read_counts
}
\description{
Quantifications of bulk RNA-seq data of liver samples from 12 male mice.
This includes only a subset of mice and genes from the experiment, for size.
}
\keyword{datasets}
