% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiked_wishart.R
\name{sparsesvd}
\alias{sparsesvd}
\title{Compute svd using sparsesvd but return a guaranteed number of singular values
since sparsesvd will truncate whenever an svd happens to small even if we set tol=0}
\usage{
sparsesvd(mat, rank)
}
\description{
Compute svd using sparsesvd but return a guaranteed number of singular values
since sparsesvd will truncate whenever an svd happens to small even if we set tol=0
}
