% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Weger18}
\alias{Weger18}
\title{Weger et al. 2018 RNA-seq of Male Mouse Liver}
\format{
\subsection{\code{Weger18}}{

A data frame with 40,614 rows and 12 columns.
Each column is one mouse sample with it's corresponding GEO identifier.
Each row is a ENSEMBL mouse gene/transcript with the quantified values for each sample.
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE77221}
}
\usage{
Weger18
}
\description{
Quantifications of bulk RNA-seq data of liver samples from 12 male mice.
This includes only the conventionally raised mice from the experiment.
Mice were collected around the clock at four hour intervals.
See PMID: 30344015 for description of experimental and sequencing setup.
See PMID: 37382061 for description of alignment and quantification using Salmon.
}
\keyword{datasets}
